+++ internal/storage/v1/elasticsearch/spanstore/reader.go
@@ -10,7 +10,6 @@ import (
 	"encoding/json"
 	"errors"
 	"fmt"
-	"strconv"
 	"time"
 
 	"github.com/olivere/elastic"
@@ -76,22 +75,6 @@ var (
 
 	nestedTagFieldList = []string{nestedTagsField, nestedProcessTagsField, nestedLogFieldsField}
 )
-
-// SpanReaderV1	is a wrapper around SpanReader
-type SpanReaderV1 struct {
-	spanReader    *SpanReader
-	spanConverter dbmodel.ToDomain
-	tracer        trace.Tracer
-}
-
-// NewSpanReaderV1 returns an instance of SpanReaderV1
-func NewSpanReaderV1(p SpanReaderParams) *SpanReaderV1 {
-	return &SpanReaderV1{
-		spanReader:    NewSpanReader(p),
-		spanConverter: dbmodel.NewToDomain(p.TagDotReplacement),
-		tracer:        p.Tracer,
-	}
-}
 
 // SpanReader can query for and load traces from ElasticSearch
 type SpanReader struct {
@@ -239,19 +222,19 @@ func timeRangeIndices(indexName, indexDateLayout string, startTime time.Time, en
 }
 
 // GetTrace takes a traceID and returns a Trace associated with that traceID
-func (s *SpanReaderV1) GetTrace(ctx context.Context, query spanstore.GetTraceParameters) (*model.Trace, error) {
-	ctx, span := s.spanReader.tracer.Start(ctx, "GetTrace")
+func (s *SpanReader) GetTrace(ctx context.Context, traceIDs map[dbmodel.TraceID]string) (map[dbmodel.TraceID][]*dbmodel.Span, error) {
+	ctx, span := s.tracer.Start(ctx, "GetTrace")
 	defer span.End()
 	currentTime := time.Now()
 	// TODO: use start time & end time in "query" struct
-	traces, err := s.multiRead(ctx, []model.TraceID{query.TraceID}, currentTime.Add(-s.spanReader.maxSpanAge), currentTime)
+	traces, err := s.MultiRead(ctx, traceIDs, currentTime.Add(-s.maxSpanAge), currentTime)
 	if err != nil {
 		return nil, es.DetailedError(err)
 	}
 	if len(traces) == 0 {
 		return nil, spanstore.ErrTraceNotFound
 	}
-	return traces[0], nil
+	return traces, nil
 }
 
 func (s *SpanReader) collectJsonSpans(esSpansRaw []*elastic.SearchHit) ([]*dbmodel.Span, error) {
@@ -262,19 +245,6 @@ func (s *SpanReader) collectJsonSpans(esSpansRaw []*elastic.SearchHit) ([]*dbmod
 			return nil, fmt.Errorf("marshalling JSON to span object failed: %w", err)
 		}
 		spans[i] = jsonSpan
-	}
-	return spans, nil
-}
-
-func (s *SpanReaderV1) collectSpans(jsonSpans []*dbmodel.Span) ([]*model.Span, error) {
-	spans := make([]*model.Span, len(jsonSpans))
-
-	for i, jsonSpan := range jsonSpans {
-		span, err := s.spanConverter.SpanToDomain(jsonSpan)
-		if err != nil {
-			return nil, fmt.Errorf("converting JSONSpan to domain Span failed: %w", err)
-		}
-		spans[i] = span
 	}
 	return spans, nil
 }
@@ -293,11 +263,6 @@ func (*SpanReader) unmarshalJSONSpan(esSpanRaw *elastic.SearchHit) (*dbmodel.Spa
 }
 
 // GetServices returns all services traced by Jaeger, ordered by frequency
-func (s *SpanReaderV1) GetServices(ctx context.Context) ([]string, error) {
-	return s.spanReader.GetServices(ctx)
-}
-
-// GetServices returns all services traced by Jaeger, ordered by frequency
 func (s *SpanReader) GetServices(ctx context.Context) ([]string, error) {
 	ctx, span := s.tracer.Start(ctx, "GetService")
 	defer span.End()
@@ -310,30 +275,6 @@ func (s *SpanReader) GetServices(ctx context.Context) ([]string, error) {
 		cfg.RolloverFrequencyAsNegativeDuration(s.serviceIndex.RolloverFrequency),
 	)
 	return s.serviceOperationStorage.getServices(ctx, jaegerIndices, s.maxDocCount)
-}
-
-// GetOperations returns all operations for a specific service traced by Jaeger
-func (s *SpanReaderV1) GetOperations(ctx context.Context, query spanstore.OperationQueryParameters) ([]spanstore.Operation, error) {
-	operations, err := s.spanReader.GetOperations(ctx, esOperationQueryParamsFromSpanStoreQueryParams(query))
-	if err != nil {
-		return nil, err
-	}
-	// TODO: https://github.com/jaegertracing/jaeger/issues/1923
-	// 	- return the operations with actual span kind that meet requirement
-	var result []spanstore.Operation
-	for _, operation := range operations {
-		result = append(result, spanstore.Operation{
-			Name: operation.Name,
-		})
-	}
-	return result, err
-}
-
-func esOperationQueryParamsFromSpanStoreQueryParams(p spanstore.OperationQueryParameters) dbmodel.OperationQueryParameters {
-	return dbmodel.OperationQueryParameters{
-		ServiceName: p.ServiceName,
-		SpanKind:    p.SpanKind,
-	}
 }
 
 // GetOperations returns all operations for a specific service traced by Jaeger
@@ -374,83 +315,28 @@ func bucketToStringArray(buckets []*elastic.AggregationBucketKeyItem) ([]string,
 	return strings, nil
 }
 
-// FindTraces retrieves traces that match the traceQuery
-func (s *SpanReaderV1) FindTraces(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) ([]*model.Trace, error) {
-	ctx, span := s.spanReader.tracer.Start(ctx, "FindTraces")
-	defer span.End()
-
-	uniqueTraceIDs, err := s.FindTraceIDs(ctx, traceQuery)
-	if err != nil {
-		return nil, es.DetailedError(err)
+func bucketToTraceIDArray(buckets []*elastic.AggregationBucketKeyItem) ([]dbmodel.TraceID, error) {
+	strings := make([]dbmodel.TraceID, len(buckets))
+	for i, keyitem := range buckets {
+		str, ok := keyitem.Key.(string)
+		if !ok {
+			return nil, errors.New("non-string key found in aggregation")
+		}
+		strings[i] = dbmodel.TraceID(str)
 	}
-	return s.multiRead(ctx, uniqueTraceIDs, traceQuery.StartTimeMin, traceQuery.StartTimeMax)
+	return strings, nil
 }
 
 // FindTraceIDs retrieves traces IDs that match the traceQuery
-func (s *SpanReaderV1) FindTraceIDs(ctx context.Context, traceQuery *spanstore.TraceQueryParameters) ([]model.TraceID, error) {
-	ctx, span := s.spanReader.tracer.Start(ctx, "FindTraceIDs")
+func (s *SpanReader) FindTraceIDs(ctx context.Context, traceQuery *dbmodel.TraceQueryParameters) ([]dbmodel.TraceID, error) {
+	ctx, span := s.tracer.Start(ctx, "FindTraceIDs")
 	defer span.End()
 
-	if err := validateQuery(traceQuery); err != nil {
-		return nil, err
-	}
 	if traceQuery.NumTraces == 0 {
 		traceQuery.NumTraces = defaultNumTraces
 	}
 
-	esTraceIDs, err := s.spanReader.FindTraceIDs(ctx, esTraceQueryParamsFromSpanStoreTraceQueryParams(traceQuery))
-	if err != nil {
-		return nil, err
-	}
-
-	return convertTraceIDsStringsToModels(esTraceIDs)
-}
-
-func esTraceQueryParamsFromSpanStoreTraceQueryParams(p *spanstore.TraceQueryParameters) *dbmodel.TraceQueryParameters {
-	return &dbmodel.TraceQueryParameters{
-		ServiceName:   p.ServiceName,
-		OperationName: p.OperationName,
-		Tags:          p.Tags,
-		StartTimeMin:  p.StartTimeMin,
-		StartTimeMax:  p.StartTimeMax,
-		DurationMin:   p.DurationMin,
-		DurationMax:   p.DurationMax,
-		NumTraces:     p.NumTraces,
-	}
-}
-
-func (s *SpanReaderV1) multiRead(ctx context.Context, traceIDs []model.TraceID, startTime, endTime time.Time) ([]*model.Trace, error) {
-	ctx, childSpan := s.tracer.Start(ctx, "multiRead")
-	defer childSpan.End()
-	traceIds := make(map[dbmodel.TraceID]string)
-	for _, traceID := range traceIDs {
-		traceIds[dbmodel.TraceID(traceID.String())] = getLegacyTraceId(traceID)
-	}
-	tracesMap, err := s.spanReader.MultiRead(ctx, traceIds, startTime, endTime)
-	if err != nil {
-		return []*model.Trace{}, err
-	}
-	var traces []*model.Trace
-	for _, jsonSpans := range tracesMap {
-		spans, err := s.collectSpans(jsonSpans)
-		if err != nil {
-			err = es.DetailedError(err)
-			logErrorToSpan(childSpan, err)
-			return nil, err
-		}
-		traces = append(traces, &model.Trace{Spans: spans})
-	}
-	return traces, nil
-}
-
-func getLegacyTraceId(traceID model.TraceID) string {
-	// https://github.com/jaegertracing/jaeger/pull/1956 added leading zeros to IDs
-	// So we need to also read IDs without leading zeros for compatibility with previously saved data.
-	// TODO remove in newer versions, added in Jaeger 1.16
-	if traceID.High == 0 {
-		return strconv.FormatUint(traceID.Low, 16)
-	}
-	return fmt.Sprintf("%x%016x", traceID.High, traceID.Low)
+	return s.findTraceIDs(ctx, traceQuery)
 }
 
 func (s *SpanReader) MultiRead(ctx context.Context, traceIds map[dbmodel.TraceID]string, startTime, endTime time.Time) (map[dbmodel.TraceID][]*dbmodel.Span, error) {
@@ -561,7 +447,7 @@ func buildTraceByIDQuery(traceID dbmodel.TraceID, legacyTraceID string) elastic.
 		elastic.NewTermQuery(traceIDField, legacyTraceID))
 }
 
-func convertTraceIDsStringsToModels(traceIDs []string) ([]model.TraceID, error) {
+func convertTraceIDsStringsToModels(traceIDs []dbmodel.TraceID) ([]model.TraceID, error) {
 	traceIDsMap := map[model.TraceID]bool{}
 	// https://github.com/jaegertracing/jaeger/pull/1956 added leading zeros to IDs
 	// So we need to also read IDs without leading zeros for compatibility with previously saved data.
@@ -570,7 +456,7 @@ func convertTraceIDsStringsToModels(traceIDs []string) ([]model.TraceID, error) 
 	// TODO remove deduping in newer versions, added in Jaeger 1.16
 	traceIDsModels := make([]model.TraceID, 0, len(traceIDs))
 	for _, ID := range traceIDs {
-		traceID, err := model.TraceIDFromString(ID)
+		traceID, err := model.TraceIDFromString(string(ID))
 		if err != nil {
 			return nil, fmt.Errorf("making traceID from string '%s' failed: %w", ID, err)
 		}
@@ -602,10 +488,10 @@ func validateQuery(p *spanstore.TraceQueryParameters) error {
 	return nil
 }
 
-func (s *SpanReader) FindTraceIDs(
+func (s *SpanReader) findTraceIDs(
 	ctx context.Context,
 	p *dbmodel.TraceQueryParameters,
-) ([]string, error) {
+) ([]dbmodel.TraceID, error) {
 	ctx, childSpan := s.tracer.Start(ctx, "findTraceIDs")
 	defer childSpan.End()
 	//  Below is the JSON body to our HTTP GET request to ElasticSearch. This function creates this.
@@ -680,11 +566,11 @@ func (s *SpanReader) FindTraceIDs(
 	searchResult, err := searchService.Do(ctx)
 	if err != nil {
 		err = es.DetailedError(err)
-		s.logger.Info("es search services failed", zap.Any("traceQuery", serviceName), zap.Error(err))
+		s.logger.Info("es search services failed", zap.Any("traceQuery", p), zap.Error(err))
 		return nil, fmt.Errorf("search services failed: %w", err)
 	}
 	if searchResult.Aggregations == nil {
-		return []string{}, nil
+		return []dbmodel.TraceID{}, nil
 	}
 	bucket, found := searchResult.Aggregations.Terms(traceIDAggregation)
 	if !found {
@@ -692,7 +578,7 @@ func (s *SpanReader) FindTraceIDs(
 	}
 
 	traceIDBuckets := bucket.Buckets
-	return bucketToStringArray(traceIDBuckets)
+	return bucketToTraceIDArray(traceIDBuckets)
 }
 
 func (s *SpanReader) buildTraceIDAggregation(numOfTraces int) elastic.Aggregation {
